% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffread.R
\name{ffread}
\alias{ffread}
\title{Title}
\usage{
ffread(file, key, value, ...)
}
\arguments{
\item{file}{character. Path to the file to read}

\item{key}{character. Column name to filter the data table}

\item{value}{character. Value to filter the data table}

\item{...}{Additional arguments to pass to data.table::fread()}
}
\description{
Title
}
